"1:35pm. Time to resume. I keep wanting to get into the Elaina novels. I'll do it today after I finish with work.
1:55pm. Next come adding paragraphs. Since I am playing around, a 4m video can take a lot longer. This stuff would actually be pretty useful if I am doing something like web design as the video noted.

It is not that great to have an 1/5 score in art. I'll want to bring that up to at least 2/5.

2:05pm. Focus me, next is text formatting.

2:10pm. Next is placing text on a path.

2:20pm. https://helpx.adobe.com/illustrator/how-to/text-basics.html

In example 4 here, why is the text on the bottom not filling in the rectangle. Instead it seems to be made up of individual shapes. What is up with that?

2:35pm. Ah, it is possible to take text and turn it into individual shapes. I see.

2:50pm. https://helpx.adobe.com/illustrator/how-to/drawing-tools-basics.html

This is next. I found the previous lecture relatively difficult as I did not have the required fonts and the default one had different shapes. But it did teach me a bit.

Now come curves.

I'll finally be able to do some of the actual art that I had in mind. So far, I really like the Illustrator as a tool.

3:05pm. I am getting lost in thought. Let me start the lessons.

3:30pm. It is really interesting what these tools are capable of. Right now I am trying to draw a sail and this tool allows me to redraw parts of the path with no trouble. This is good, since drawing with the mouse is really hard.

3:40pm. This chapter is less boring than the other ones since I have to stuggle a bit to get the lines to match up. I want to know more.

4pm. I've made the right decision to study this. If I just tried fiddling with the program on my own, it would take me a lot of time to figure all this out. It is best to do it up front.

Why am I doing this? Why did all of a sudden I decide to cultivate art skills?

Simply because GANss and VAEs and their combination is going to be a big deal in the future. NNs are a particularly big deal for art related work and if I restricted myself to RL, I would be quite poor in my capability to advantage of their full range of features.

RL is ultimately just a small part of the cake, and the generative stuff underneath is what is going to power the sample efficiency gains. So I might as well get familiar with it.

I need to change my style of doing things. I do not care if selling entertainment is what makes me money, I am going to stop playing life with restrictions.

4:40pm. https://helpx.adobe.com/illustrator/how-to/drawing-tools-basics.html

I find myself spending a lot of time on the last example of this and I am still not done. Why am I having so much trouble joining the paths. It is not working all like in the video. The behavior of it is so strange and it seems to have trouble recognizing one of the edges properly.

4:45pm. Ok, I figured it out. It seems that the curve is already duplicated stealthily, so copying it causes it to be copied twice. I only noticed this because of the curve tool.

Let me just finish the video and I'll do the chores.

4:50pm. Ok, I am satisfied. Let me do the chores.

https://helpx.adobe.com/illustrator/how-to/ai-layers-basics.html

The next are the layers basics.

5:10pm. Done with chores. I am looking at the old example again. It is as if somebody played a prank by connecting a line to the other end.

Well, let me move on. Amazing how much time these ultra basic tutorials are taking me. But I got into this last section.

5:20pm. Focus me, let me go through this section and I'll call it a day.

Leave the Umineko thread for later. I read ep1-6 ages ago, but not the later. I've forgotten most of it by now.

5:30pm. Ah, I see. I can use the layers panel to rearange things more precisely. Nice.

5:35pm. Focus me, two more videos and then I am done.

5:45pm. One more and I am done.

5:55pm. https://helpx.adobe.com/illustrator/tutorials.html
https://helpx.adobe.com/illustrator/how-to/artboards-basics.html

I'll leave this one for tomorrow.

It seems there are only 3 lessons left in the Get Started course.

6pm. Hmmm, thanks to what I have in my wallet (gifts from relatives I never spend) and the Spiral donations, I can just about afford to buy a tablet if I feel like it. Right now, I am still having some doubts on how long I'll stick with this, but if I can get good with vector art and feel that it's not enough, I'll seriously consider it.

6:05pm. GANs...right now I do not want to go anywhere near those things. I am so asspained from poker, that I never want to touch NNs with current hardware again. I can't run any of the bigger GANs easily, so that will send me down a rabbit hole of trying out hacks much like with Holdem. It would be a huge waste of time.

Really, I can't fault myself from wanting to develop my personal skills instead of braving that hellish gauntlet again.

AI chips would revive my interest, but now I am completely defeated. At some point one need to learn to throw in the towel. I can't do the same thing my whole life and hope that it works.

It is fine to return to Simulacrum every couple of years.

Since I would not feel good about only just writing again, I have to do this to prove my sincerity in the quest for self improvement. Everything I do should have a new basis behind it. That is the way to stay fresh. Just churning out text is to invite decay."