"9:20am. I am up.
9:25am. https://www.reddit.com/r/MachineLearning/comments/mif2i3/r_on_the_role_of_planning_in_modelbased_deep/

An interesting paper got posted on the RL sub.

9:30am. Let me chill and I will start. My body is starting to feel better. When bad posture starts killing your back, the thing to do is to change your habits and move often, both while being seated and in bed. This is why I often sway while in this seat.

https://arxiv.org/abs/2011.04021
On the role of planning in model-based deep reinforcement learning

Let me read this paper. I'll chill a bit first and then move forward. Yesterday I got Umineko 5-8, but will leave them for later. I think I'll start from 5 since I forgot everything anyway. I must have been a decade since I read the fan translation.

10:20am. Let me start. But before that let me write down the intro poem for Heaven's Key. I thought about it during the night.

$$$

Through a desert a traveller walks
Even as his blood boils into vapor
As his flesh decays and rots
As his bones crumble into dust
He continues moving

As his body dies, he leaves the unneeded behind
And continues moving forward

His soul blazing, he traces out heavy steps
Seeking his goal

Far above
The Herald of the 3 Gods' voice resounds:
Break the Limits! Ignite the Singularity! Pursue Omnipotence!

- Intro to Heaven's Key

$$$

This is a scene from the life of one of the Inspired captured in a poem. The most important thing is not rhyme or cleverness, but capturing the image. It has some wonderful contrast between the harshness of life and the determination of the seeker.

10:40am. This is what the Inspired are.

Forget the Path of Conqueror. I've been destined to step on the Path Of Illusionist from the start. Of course I cannot win so easily. If just making use of some RL algorithm was enough, poker would have long been conquered.

10:50am. Let me start my drawing journey. I'll want to get to at least 2/5 in both graphics and audio before looking into GANs. The goal of the illusionist is to automate as much of his creative process as he can.

Some artists are afraid of machines taking over, but replacing my own effort with theirs is exactly the goal of this path. The less I write, the less I draw and the less I compose, the more successful I can consider myself to be.

Rather than trying to push RL to conquer games, this path aims to master generative models. There is definitely a tie into PPLs here whereas the conqueror's path only has a weak connection.

Given a goal, there are many ways to reach it, and some will be better than others. I chose incorrectly when I picked poker.

On the other hand, if I can master the generative aspects of intelligence, conquering games will be much, much easier. In humans the generative aspects underlie all of their power. The brain is certainly not doing end-to-end RL.

10:55am. Let me start. Where was I yesterday?

11am. Focus me.

https://helpx.adobe.com/illustrator/tutorials.html
https://helpx.adobe.com/illustrator/how-to/artboards-basics.html

Done with the first two. Since this is just about artboards it is easy as expected. Let me move to the last one.

11:35am. The next are images and effects.

https://helpx.adobe.com/illustrator/how-to/image-effects-basics.html

12pm. Had to take a break. Let me resume.

12:05pm. Done with the first lecture. Let me stop here.

The way it is shaping up in my head, Heaven's Key is going to be the best Simulacrum arc yet. In 2014 it was just a follow up to what I had before, but now I am starting to see its true potential.

Sigh though, can I really expect it to sell? Wouldn't it be nice if it got some attention instead of getting passed over on the global marketplace? That is what will probably happen. It is a tough world out there.

It is like going through a desert. I do not know when the end will come, I can only keep going forward.

When the AI chips come out for sale, I will definitely make up my mind and get a job if I haven't done that yet. The easiest route will be to get back to Z and accept his terms of 2.5k/month assuming he is not full by that time. If he raises it on his own initiative, then I'll be able to trust him. If not, I'll leave after a couple of months after having mastered Julia macrology and PPL construction.

I really do not feel like doing any other jobs than this. Everything else would require me to change my interests.

Though in a year, I will have an expanded skillset. Especially if I master the use of GANs, I might have the opportunity to do paid creative work. It depends on how it goes.

12:15pm. Sigh, I want money, but I sure do not want to work for others. Sigh. Well, I'll get over it at some point. If I did not need to cultivate, things would be easier. Ordinary people can do whatever falls into their lap, but I myself need to guide my development."