"9:45pm. https://youtu.be/44HBr8O-kPk?t=404
I've just realized that the guy was using Photoshop. Take a look at the top left corner. Visually PS and AI looked so similar that it never occurred to me to actually check what program he is using. I just assumed it was the Illustrator. That settles that mystery. I'll have to go for freeform gradients, I can't just blend colors like the dude did. Well, actually, it might be possible if I make a cutout of the surounding layer and blend that instead. But the actual gradient tools would be more ergonomic.

Freeform gradients are a wonderful thing so I should use them.

That settles one mystery.

9:50pm. Tomorrow I will watch a few more videos and then get cracking. Apart from not really knowing how to combine different artworks into something unique - so far I've only seen traced portraits, it bothers me that I only have two tracks composed in my head. They caught my attention and I want to create them, but I'll need more. At least a dozen or two.

10:05pm. This is a bit fun. I've figured out the third one. I was trying to figure out something to represent transcendence, but nothing is coming to me. It is too nebulous as a concept, and I can't associate a melody to it.

Even though it was by accident, I figured out the Violence track.

So far I have Boredom (At School), Doom (For NPCs) and (Transcendent) Violence.

It is really coming together well in my mind. This is really a new experience for me, if I was just my old self doing just writing I would never have reached this level. I would have just languished where I was before.

There is power in attempting to what was thought to be beyond one's grasp.

10:10pm. This could turn out really well.

The Boredom track especially will be very catchy. In fact the whole school sequences will give a voice to a feeling never ellucidated before. I think just the intro being done properly will grasp the hearts of those who play it.

Yeah, the ingredients are all there to create a roaring flame that will lift me higher in my air baloon, but it is still up to luck. That is the nature of the market. To ensure the success of the game, I must never indulge. This time I'll be a pro from start to finish.

10:20pm. I am actually having a lot of fun listening to Boredom track in my mind. The only thing I feel towards it is longing.

10:40pm. Maybe longing is what should be the motif of the Transcendence song. (Longing For) Transcendence.

https://www.reddit.com/r/MachineLearning/comments/q2u2kx/d_paper_explained_grokking_generalization_beyond/

Let me take a look at this. After that I'll watch slime.

I've played Vigil: Longest Night for a while, but its gameplay is not that good. It's artwork is well done, but that is it."