"10:05am. Let me chill a bit and then I will start.
10:40am. Let me start. I forgot that I had those two vols of Pandora in the Crimson Shell yesterday. I'll leave them for later.

The illusionist path, go!

It connects to the conqueror's path when it comes to games, but for poker for example, in the illusionist path, I'd actually deal with UI automation first. I read that a good agent, but no interface is worth a fortune and the reverse is worth squat, but that is wrong. I should have gone the opposite route.

Having the agent learn to immitate my moves is not that different from having it learn to 3d model or program like me. I should have dealt with self supervised aspects of learning first. RL is a dead end.

I need a hint on what the principles of intelligence are. Then I will be able to pursue AI for real again. I do not want to waste any more time with ML papers.

10:45am. I did so much useless stuff in the past 6.5 years. If I put in as much effort into learning to draw as I did in studying random stuff in 2019 and 2020, I'll master it without a problem.

Now where was I with the course?

https://youtu.be/g8683RF1COo?list=PLn3ukorJv4vs_eSJUQPxBRaDS8PrVmIri
Complete Beginners Guide to Blender 2.8 | Free course | Part 6 | Rendering and Atmosphere

I am done with the man and monster and it is time to take care of the street, lighting and fog.

10:50am. This will be good. The last 6.5 years have been so unprofitable apart from building up my programming skills. In the next 6.5, maybe I won't make AGI, but I'll have something to show for it at least. When grand goals aren't working, it is good to have some humility and stay out of trouble.

10:55am. Let me start with the pavement and the lamp posts.

11:35am. https://youtu.be/g8683RF1COo?list=PLn3ukorJv4vs_eSJUQPxBRaDS8PrVmIri&t=482

I am also getting flicker and there are some weird cutoffs due to the way light is rendered.

12:30pm. That is is for the beginner's course. Let me watch his other 3 short vids and then I'll stop for breakfast.

12:50pm. I'll leave paid resources for much later.

https://www.youtube.com/playlist?list=PLn3ukorJv4vvJM7tvjet4PP-LVjJx13oB
learn sculpting in blender 2.8

Here is a sculpting course by Grant. Maybe I'll go for that next. I'll watch another of FlyCat's vids first as a reward for finishing the first course. Right now I am about mid 1/5 in Blender. I am not an absolute beginner anymore, but I still do not know most about it.

How about this? FlyCat's ninja girl video, the manga in Blender Youtube tutorials, and then I'll move to the scuplting course. Then I'll try making a 3d hand. If I can't I'll do more free courses.

12:55pm. Doing it gradually like this is the way to build skill. Maybe I'll be able to actually do something on my own in a few months."