"10:10am. I am just chilling. Let me start. In honor of his course, let me take a look at the live stream.
https://youtu.be/dc3SyGNXINM?list=PLdFBjtsWmM96jvwSbyMT5vBFmbgt59Ftk
Sculpting Thor | Live Stream Sculpting in Blender

10:20am. https://www.youtube.com/watch?v=mnnMFJ3iWH0
New Features in Blender 2.93 LTS in LESS than 5 minutes

Let me take a look at this first. That this vid has 260k views is telling about how popular Blender is.

https://youtu.be/mnnMFJ3iWH0?t=134

Has a bunch of 2d stuff.

https://youtu.be/dc3SyGNXINM?list=PLdFBjtsWmM96jvwSbyMT5vBFmbgt59Ftk&t=426

Let me just watch this for now.

10:50am. https://youtu.be/dc3SyGNXINM?list=PLdFBjtsWmM96jvwSbyMT5vBFmbgt59Ftk&t=892

Now I am listening to him talk to people in the chat. Why is this vid in the sculpting playlist?

Let me move on to Grant's course. I do not feel like hunting down a reference of Thor and he did not display it.

https://www.youtube.com/playlist?list=PLn3ukorJv4vvJM7tvjet4PP-LVjJx13oB
learn sculpting in blender 2.8

Here is the playlist by Grant. Let me try it again.

11:10am. https://youtu.be/lKY2FIy60nc?list=PLn3ukorJv4vvJM7tvjet4PP-LVjJx13oB&t=195

I do not like this process too much as it is really difficult to see what is happening when deforming a sherical object. But I'll roll with it. It have been easier to flatten certain sides a little before doing anything else.

12pm. I stopped following along after the first half, but I got the gist of it. Ok. If it is just sculpting I feel it might be worth it to try sculpting a hand. But I want to round out my beginner's education first. I am going to watch the rest of Pinkbeard's tutorials and then try pursuing my own projects after that. Just the little bit I did I find satisfactory since it does look like a face. This never happened to me in drawing.

In order to improve I need to sculpt based on a reference.

12:05pm. I need to improve my grasp of Blender just a bit more and then I'll be able to do that.

Let me have breakfast here."