"2pm. Let me resume. What is next on the list? The basic tutorials by Pinkbeard.
https://www.youtube.com/c/SirPinkBeard/playlists

2:05pm. Right now I am thinking about the next Simulacrum arc after Heaven's Key and am having a bit too much fun. I need to focus.

2:15pm. https://youtu.be/fuddZUv7MzA?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U

This sounds pretty useful. Everything here is new. There is so much to cover. Just think how poorer my life would have been if I had not decided to branch out.

2:30pm. https://youtu.be/fuddZUv7MzA?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U&t=332

All of these would be good for terrain. This one in particular reminds me of the level brush. It might be easier to handle.

2:45pm. Focus me, let me move to the next video. Leave the threads for later.

3:05pm. https://youtu.be/S7rJ4lolbus?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U

Let me take a short break here.

3:30pm. Let me resume. Time for loop cuts. It is going to take me a while to get through all of these playlists, but after I am done I'll have a decent Blender starting education. After that I'll be able to try sculpting hands and moving on to immitating FlyCat.

https://youtu.be/S7rJ4lolbus?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U&t=83

I thought I was familiar with loop cut, but I had no idea you could use the wheel to select the amount.

I managed to crash Blender while playing with the edge cut tool.

4pm. https://youtu.be/Q-NjY7I3Zcw?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U&t=424

I now understand why the default sculpting setup is a subdivided cube.

https://youtu.be/c5QDKMwvbkw?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U&t=7
Creating a Basic Sword [Blender 2.8]

Since I am starting to fall asleep it might be better to give this a try.

https://youtu.be/c5QDKMwvbkw?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U&t=775

This is a really interesting way of creating it. I am going to have to try this myself.

I am falling asleep just watching him talk about tools, and a light exercise should serve to sink the lesson in. I'll the mirror modifier though to save me time.

5:10pm. This did not come out as planned. The mirror modifier actually made things more difficult instead of saving me work.

Never mind repeating this right now. Let me watch the video again and then I'll move on.

6:10pm. I am too obsessed with getting the symmetry for that sword right. Let me just stop here with it. Let me move to the next lecture.

6:35pm. I finally got the sword right. Let me see the trick he used to level the tip.

Ah, scale + z + 0.

I get it.

7:25pm. Done with lunch.

https://youtu.be/Op1PHiNGNao?list=PLdFBjtsWmM96iY6Zz67wpv_WhK_1h-11U
The Smooth and Randomize Tools [Blender 2.8]

I'll leave this for tomorrow, let me stop here for the day. I spent a lot of time on that sword, but that is a good thing. It shows how immersed I am in the material. This kind of immersion will no doubt translate into actual skill down the road."