"2:50pm. Done with the second round of chores. Let me cool down just a bit.
3pm. Let me start. Focus me.

I am done with the scene and he is going to demo the rigid body sim. That should be interesting.

3:10pm. This is quite cool. It took me a while to realize why the movement is so weird, but the stone blocks feel very light, plus the gravity seems to be light as well.

3:15pm. Ah, I think I get why the movement is so unusual. It might have to do with the pivot points.

3:30pm. Done with part 1. Next is part 2.

3:40pm. Got a reply from Pinkbeard.

> You need to Extrude out a vertex to act as the hip joint. That's why I can grab that vertex on the right at that timestamp.

Extrude? Let me give it a try.

4pm. Yeah, that was it. Though it looks different than for him than it does for me.

4:05pm. Nevermind that for now. Let me keep following the CG Fast course, since I am on it. I'll return to Pinkbeard's stuff later.

4:25pm. Let me do it for 20m more and I'll take a break.

5:10pm. https://youtu.be/saxpjudV_88?t=1192

I am playing with the Bevel modifier, but it is making the shading look weird.

...I realized why - it is because it does beveling on low angle edges. Setting it to 80 degrees fixes it.

5:20pm. https://youtu.be/saxpjudV_88?t=1387

Time for lunch. Let me pause here.

5:55pm. Let me resume. Let me just watch the next part and then I will close for the day.

https://youtu.be/saxpjudV_88?t=1572

I've been thinking about the previous sword and how it might have been a good idea to do the hilt jewel as a separate object rather than extruding everything.

https://youtu.be/saxpjudV_88?t=1388

Ok, let me stop here for the day. I'll resume this tomorrow. Right now I want to play Salt and Sanctuary. I gave it a try yesterday. It feels like it is by the Zombie Smasher guy going by the art style.

It is really my kind of game, it is amazing how well it captures the Dark Souls feel. I am going to have to look at the guide for how to play it. I realized that I had no idea how to build up the mage yesterday and I messed up the character. I would've put points into staves, but the merchant for it was hidden behind a secret wall."