"10:30am. I went to bed past 1am yesterday. The gaming session was great, but I think I made a mistake going with a mage as I am just spamming lightning everywhere. I should switch to melee for more interesting battles.
Salt and Sanctuary is great all the way. I am not sure if I'd place it higher than Ender Lilies, but it is better than Hollow Knight which is better than Bloodstained: Ritual of the Farming.

Let me read Frieren and Kenja no Deshi and I'll resume the CG course.

10:45am. I really hate how the Kenja is getting translations to parts that have already been done years ago, while the raws are getting translated at a tepid pace. Its scanlators are morons.

Forget that, let me start. CG Fast Track go. Time to deal with that sword. I'll follow the video exactly. It is really cool how it spent a few minutes telling me the proportions and the coordinates of all the basic shapes.

11am. https://youtu.be/saxpjudV_88?t=1610

Got the primite shape set up.

I'll make it my goal to go through part 2 and part 3 for today. I should be able to go through part 2 at least.

11:05am. Got distracted. Focus me.

https://youtu.be/saxpjudV_88?t=1697

I am learning a lot. I had no idea that scale can and should be applied like this.

12:10pm. https://youtu.be/saxpjudV_88?t=2434

The way he hacks off one side in wireframe mode is good. It would not have occurred to me. At the rate I am going, it will take me an entire day just for this single lecture. There is no rushing this.

I am quite into it. This is good for learning, but not great for going through the material quickly.

https://www.reddit.com/r/MachineLearning/comments/qbq5bs/r_an_introduction_to_probabilistic_programming/

Oh, this PPL book got updated. Let me just do a quick skim.

Oh, I see it has 90 more pages than last time. There are a few more sections in the TOC.

12:15pm. Let me move on. PPLs are both the past and the future, but not the present for me.

If I wanted to be monster hunter, I'd want to trade my kills for money. Imagine if monster parts had zero value and I had to be a chimney sweeper to buy equipment with which to monster hunt. I'd say, fuck that shit. Monsters can eat humans for all I care. That will teach them a lesson. Right now, that is how I feel about ML in general. It should pay for itself instead of wanting me to nurse it into shape. I am not going to get a job just to get an AI chip. The path itself should give me the resources to pursue it.

AI is important, but I do not want such a worthless, unprofitable path.

With CG, I'll be able to use it in my own game dev work when I have a decent level of skill. Same with musical composition.

12:20pm. Focus me.

https://youtu.be/saxpjudV_88?t=2434

This is where I am currently. Let me watch this part.

12:25pm. https://youtu.be/saxpjudV_88?t=2489

He is talking about the pinching here. I had so much trouble in the past with this, and I could not figure out how to deal with it.

12:50pm. I figured out how to create a perspective based on the bottom face's normals. Then I used that to extrude the last side of the guard. Nice. Now I can be sure they are perfectly parallel.

1:40pm. https://youtu.be/saxpjudV_88?t=3086

Oh whops, I still did not have breakfast. Let me stop here for a bit."