"12:25pm. I think I should just go to bed on time today. Last night I could not fall asleep, and only woke up just now. The 1am SS gaming sessions aren't helping either.
12:30pm. But it fine. I needed time to think. Maybe I am wrong about various things.

12:50pm. Let me do the chores here.

1:55pm. I messed up my sleep, so I am not going to be able to much today. Let me rant a little.

During the night I've been reviewing where I went so wrong. Especially with deep learning.

A certain thought is trying to come to the suface.

Maybe the reason why I cannot get results is because I am in the wrong problem domain. Why can nature do it and not I?

For nature, its problem was how to get independent neural units to coordinate effectively. It boils down to being an async communication challenge. While I, and the community at large is only interested in optimizing a single model. This is a consequence of the hardware that we have.

2pm. Maybe me being a loser is a consequence of hanging out with the wrong crowd so to speak. If I tried to challenge the async commm problem directly, I would not be able to do anything.

But the nature being able to do it is partly a consequence of it setting up the problem so it can arrive at the solution.

2:20pm. In the past 6.5 where I went right is where I cultivated my programming skills. Where I went wrong is cultivating my ML and RL skills.

AI chips are the future, but looking at the past 5,10 and 20 years, my spirit has been damanged by elaborate long term plans. To pursue my main path, I should be getting better at ML, but no. Not like this. I do not want to spend my limited effort and energy on something as unprofitable as this.

With my current ML skills all I can do is get some code monkey data wrangling job. The data science posts might sound sexy, but curating datasets is as tedious as one can imagine work being. It is far as it can be from cultivating agents.

2:25pm. I am thinking about the Simulacrum: Hell's Prototype game. Take Doom Eternal for example. The game is good, but like all modern games, it's scale is quite small and rates poorly on the environmental damage score.

I want to see buildings sheared, mountains razed and seas boiled. I want to see cities going up in nuclear fire. I do not want to be some badass that fight the armies of hell one demon at a time.

Instead the next step in gaming should be to go higher and bigger.

I want to fight with weapons of mass destructions against tens of thousands and millions of monsters.

2:30pm. AI chips come into the story at this point. With the CPU, the number of agents I can simulate is quite small. 10-100 is the limit. But with AI chips I can begin the practice of making them asynchronous.

Could I have a game with 100k monsters on the same level with one of those AI chips? It is possible.

And ML also comes into the picture at this point. With deep learning it is absolutely impossible to do the following, but with the right memory system, it would be possible to use it to approximately memorize the environment. This would allow me to create a dynamic memory system for the game resulting in a very lifelike feel.

At the scale I am imagining, most of the game content would need to be proceduraly generated.

When you have just static assets things are easy, but in a lifelike game things would move around. If a building blows up and scatters debris everywhere I want the environment to be in the same state it was left in. Storing the exact state of a vast amount of generated content would be impossible, so I need a memory system to approximate the game state.

2:40pm. Such a memory system would solve all my issues. It would be useful for RL, for attention, for storing past experiences.

The way to figure it out might not be to do anything special.

I should simply put myself in the right domain and try to make use of the AI chip hardware to the fullest extent possible. It says AI on the surface, but it does massive parallelism on the inside. Instead of being despodent, I should thank deep learning for spearheading their development.

2:50pm. I need a miracle. I can't stand this powerlessness. I am surrounded by intelligence everywhere, yet have no ability to actually understand it. Just how stupid I am? I completely feel like an NPC.

I can't achieve anything with my current level of power. I want so much, but none of what I want is achievable.

3pm. I need to change. If a mmemory system is what I need to develop, and nature's steps are what I need to retrace, then I should not be doing poker.

Instead I cultivate my ability to do massively parallel programming of the kind those AI chips will allow.

And this time around I should learn from my past mistakes and do it in a way that is immediately profitable.

3pm. I won't compromise anymore. If I am to be friends with people, I want them to be on the same page as me.

If I am doing programming, no more years long isolated projects that might only benefit me years down the road.

What I am going to do now, I could have done years and decades ago had I the right mindset.

3:05pm. If I had a memory system instead of that huge mess that deep learning is, then I could use it in the same way as any other data structure. Want to store that compute generate building? Simply put it in.

Deep learning on the other hand is so hard to deal with. Just for poker - a very simple game, I had to significantly architect it just to get the data pipeline going. Being able to just pass a message to some system to store or fetch would be so much easier.

There is huge incentive to make the mind as the swarm.

3:10pm. I won't accept that I was right. Compared to somebody who took up game programming and made immediate profits and steady improvements, I am just a tryhard. Nature can do it and it has an IQ of 0 and I can't. What does that say about me?

3:15pm. I need to move beyond deep learning. I might not know how, but the first step is to dream about it. Forget GANs. Imagine a system that can memorize 3d models of the kind I am doing in Blender. Imagine a system that can predict the results of physics simulations. That is how I will know that I am on the right track.

That is the most basic of the baics? Create a simulator. And then replace its data structures with neural memory systems. It shuold just be a trivial change of the types being used.

3:20pm. Only with such an approach I can say that I am not doing moon shots.

Only by focusing all my attention on the simulator instead of the specifics of the learning system can I claim that I am pursuing the future with my own power.

3:20pm. The commoners aren't wrong in their approach. Going back to fundamentals is not working on PPLs. It is working on simulators and finding ways to profit off them. I do not need any ML or RL for that. When things are ready, the AI chips, the resources and whatever I need will come to me."