"4:25pm. Done with the bath. I think I am too rough on RL. Deep learning aside, I feel that I've learned from quite a bit from studying the counterfactual regret algorithms. It took me a long time to understand their operations - I only did it by contrasting it with MCTS, but now I think I completely understand their function. Fully understanding Bayesian reasoning in just the tabular case is not nothing. I sense that it will be useful at some point. Deep learning on the other hand always requires one's full attention and is not roboust, so it will be disposed of eventually.
I need the memory system to really benefit from RL. Absolutely.

Since none of the PPLs fully encompass CFR's functionality, I am not wrong in skipping them. I am justified in doing so on the basis of that.

4:30pm. Being an artist is not a bad fit for me. I go from working for a boss to working for the market at large.

It is a path that can be made to work.

Even the relative poverty of the average artists cannot compare to the poverty I've been enduring for half my life now. Even the unsucessful artists are vastly better off than me.

If it is something like a year, that should be more than long enough to get my visual and music skills to a respectable level. After that I'll be set to open shop.

4:35pm. The truth is I enjoyed doing the poker + RL stuff. If it is game programming I want to do more of that, but this time in a profitable fashion.

4:40pm. Let me resume the fast track course. I am going to skip the gaming session for today. This art stuff, instead of considering it like programming, I should treat it more like a game.

Being a indie game dev would be ideal. I'll be able to do what I want without compromises. Forget getting rich. That might happen or it might not. If I can make a win on the commoner path, that will serve as a stable foundation.

Now where was I yesterday?

5:05pm. https://youtu.be/JUhq6cHTuOE?t=93

Right now I am wondering why HDR image does not show up on layout view.

5:10pm. That video is pretty good.

https://youtu.be/saxpjudV_88?t=5997

Let me resume from here.

5:20pm. https://youtu.be/saxpjudV_88

Did not expect it to put it on the roughness property.

6:15pm. Got the textures slapped. This stuff is definitely more elaborate than it would seem at first glance.

https://youtu.be/saxpjudV_88?t=6908

My own sword is nice and glowy. Only 15m in part 2 are left.

6:20pm. Ah, the pommel is subdivided. How annoying.

https://youtu.be/saxpjudV_88?t=7079

But his solution for selecting all of this is wonderful.

6:30pm. I can't get the angles right for the pommel. Let me see what he will do.

6:50pm. Had take a break. My mom's illness is really messing up the lunch schedule. Sigh.

I've realized something while thinking during the break. If I go the job route, I might get something in the 5-10k range, but if I step on that path then trying to switch to what I am doing now means I'd have to endure years of poverty. No way is that going to happen. It is much easier to go from 0 to 100$ a month than from 10k to 100$.

There is a very good chance if I fall into the job trap, I'll drop out of the race for the Singularity.

7pm. Focus me. Instead of browing /ic/ let me get back to the video. There are only 10m left in part 2.

7:05pm. Oh it is done. I got it wrong. I am not sure I feel like part 3 here. Let me start it up.

7:45pm. Done with lunch. Let me stop here for the day.

https://youtu.be/y0DMMwwdBaY?t=79

I barely did much apart from putting the resources in the right folder. For some reason they are 600mb big."